refactor and organize
* refactor for organization
* remove unused grdb demo functionality
